+++ usaspending_api/accounts/serializers.py
@@ -1,10 +1,10 @@
 from rest_framework import serializers
 
-from usaspending_api.accounts.models import TreasuryAppropriationAccount
-from usaspending_api.accounts.models import AppropriationAccountBalances, FederalAccount
-from usaspending_api.financial_activities.serializers import FinancialAccountsByProgramActivityObjectClassSerializer
+from usaspending_api.accounts.models import (
+    AppropriationAccountBalances, FederalAccount, TreasuryAppropriationAccount)
 from usaspending_api.common.serializers import LimitableSerializer
-from usaspending_api.references.serializers import RefProgramActivityBriefSerializer, ObjectClassBriefSerializer
+from usaspending_api.references.serializers import (
+    RefProgramActivityBriefSerializer, ObjectClassBriefSerializer)
 
 
 class AppropriationAccountBalancesSerializer(LimitableSerializer):
@@ -40,10 +40,6 @@ class TreasuryAppropriationAccountSerializer(LimitableSerializer):
             },
             "account_balances": {
                 "class": AppropriationAccountBalancesSerializer,
-                "kwargs": {"read_only": True, "many": True}
-            },
-            "program_balances": {
-                "class": FinancialAccountsByProgramActivityObjectClassSerializer,
                 "kwargs": {"read_only": True, "many": True}
             },
             "program_activities": {
