+++ usaspending_api/accounts/views/tas.py
@@ -1,7 +1,9 @@
 from usaspending_api.accounts.models import TreasuryAppropriationAccount, AppropriationAccountBalances
-from usaspending_api.accounts.serializers import TreasuryAppropriationAccountSerializer, AppropriationAccountBalancesSerializer
+from usaspending_api.accounts.serializers import (
+    AppropriationAccountBalancesSerializer,
+    FinancialAccountsByProgramActivityObjectClassSerializer,
+    TreasuryAppropriationAccountSerializer)
 from usaspending_api.financial_activities.models import FinancialAccountsByProgramActivityObjectClass
-from usaspending_api.financial_activities.serializers import FinancialAccountsByProgramActivityObjectClassSerializer
 from usaspending_api.common.mixins import FilterQuerysetMixin, AggregateQuerysetMixin
 from usaspending_api.common.views import DetailViewSet, AutocompleteView
 from usaspending_api.common.mixins import SuperLoggingMixin
@@ -94,3 +96,40 @@ class TASCategoryAggregate(SuperLoggingMixin,
         queryset = self.aggregate(self.request, queryset=queryset)
         queryset = self.order_records(self.request, queryset=queryset)
         return queryset
+
+
+class TASCategoryQuarterList(SuperLoggingMixin,
+                             FilterQuerysetMixin,
+                             DetailViewSet):
+    """
+    Handle requests for quarterly financial data by appropriationappropriation
+    account (tas), program activity, and object class.
+    """
+    serializer_class = FinancialAccountsByProgramActivityObjectClassSerializer
+
+    def get_queryset(self):
+        queryset = FinancialAccountsByProgramActivityObjectClass.objects.all()
+        queryset = self.serializer_class.setup_eager_loading(queryset)
+        filtered_queryset = self.filter_records(self.request, queryset=queryset)
+        ordered_queryset = self.order_records(self.request, queryset=filtered_queryset)
+        return ordered_queryset
+
+
+class FinancialAccountsByProgramActivityObjectClassListViewSet(
+        SuperLoggingMixin,
+        FilterQuerysetMixin,
+        DetailViewSet):
+    """
+    Handles requests for financial account data grouped by program
+    activity and object class.
+    """
+
+    serializer_class = FinancialAccountsByProgramActivityObjectClassSerializer
+
+    def get_queryset(self):
+        """Return the view's queryset."""
+        queryset = FinancialAccountsByProgramActivityObjectClass.objects.all()
+        queryset = self.serializer_class.setup_eager_loading(queryset)
+        filtered_queryset = self.filter_records(self.request, queryset=queryset)
+        ordered_queryset = self.order_records(self.request, queryset=filtered_queryset)
+        return ordered_queryset
